core: provide scall_save_panic_stack()
Provides an architecture specific function scall_save_panic_stack() in
core/arch/arm/kernel/arch_scall.c. This function and its helpers are
extracted from core/arch/arm/tee/arch_svc.c.

Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
